+++ unit_tests/simple_write.cpp
@@ -4,7 +4,7 @@
 #include <unistd.h>
 #include <mpi.h>
 
-void initialize_data(int* data, int num_elements, int num_writes, int rank) {
+void initialize_data(int* data, long int num_elements, long int num_writes, int rank) {
 	for (int k = 0; k < num_writes; ++k) {
 		for (int i = 0; i < num_elements; ++i) {
 			data[i + k * num_elements] = i + k * num_elements + rank;
@@ -14,15 +14,15 @@ void initialize_data(int* data, int num_elements, int num_writes, int rank) {
 
 int main (int argc, char** argv) {
 	int rank;
-	int num_elements, num_writes;
+	long int num_elements, num_writes;
 	MPI_Init(&argc, &argv);
 	if (argc != 3) {
 		std::cout << "simple read input error" << std::endl;
 		MPI_Finalize();
 		return 0;
 	}
-	num_elements = std::atoi(argv[1]);
-	num_writes = std::atoi(argv[2]);
+	num_elements = std::atol(argv[1]);
+	num_writes = std::atol(argv[2]);
 	MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 	int* data = new int[num_elements * num_writes];
 	initialize_data(data, num_elements, num_writes, rank);
